+++ CHANGELOG.md
@@ -14,6 +14,7 @@ non-deterministic and rely on pulling data from the network.
 - Simplify machine-service placement. For example, deploying a service to a
 floating IP is now expressed as `myService.placeOn({floatingIp: '8.8.8.8'})`.
 - Remove `Service.connect`. Only `Service.allowFrom` can be used from now on.
+- Restart containers if their hostname changes.
 
 Release 0.2.0
 -------------